Revalidate Market List when Market is Created
