+++ packages/markets/components/CreateMarketForm.stories.ts
@@ -1,0 +1,12 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { CreateMarketForm } from './CreateMarketForm'
+
+const meta = {
+  component: CreateMarketForm,
+  tags: ['autodocs'],
+} satisfies Meta<typeof CreateMarketForm>
+
+export default meta
+type Story = StoryObj<typeof meta>
+
+export const Default: Story = {}
